# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "signum API.",
        "title": "signum",
        "contact": {},
        "version": "dev"
    },
    "basePath": "/",
    "paths": {
        "/api/badges": {
            "post": {
                "description": "Stores a badge definition and returns its id and token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Badges"
                ],
                "summary": "Create a badge",
                "parameters": [
                    {
                        "description": "Create Badeg Request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateBadgeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/CreateBadgeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/badges/live": {
            "get": {
                "description": "Renders an SVG badge for the provided parameters.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Badges"
                ],
                "summary": "Render a live badge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Left-hand subject text",
                        "name": "subject",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Right-hand status text",
                        "name": "status",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Badge color (named or hex)",
                        "name": "color",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Badge style (flat, flat-square, plastic). Default: flat",
                        "name": "style",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SVG image",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/badges/{id}": {
            "get": {
                "description": "Returns an SVG badge for the stored definition, with optional query overrides.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Badges"
                ],
                "summary": "Render a stored badge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Badge ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Override subject",
                        "name": "subject",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Override status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Override color",
                        "name": "color",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Override style",
                        "name": "style",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SVG image",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes the stored badge definition.",
                "tags": [
                    "Badges"
                ],
                "summary": "Delete a badge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Badge ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates one or more fields in the stored badge definition.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Badges"
                ],
                "summary": "Patch a badge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Badge ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Patch Badge request",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/PatchBadgeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Badge"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/badges/{id}/meta": {
            "get": {
                "description": "Returns the stored badge fields without the token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Badges"
                ],
                "summary": "Read badge metadata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Badge ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Badge"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Badge": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "style": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "CreateBadgeRequest": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "style": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                }
            }
        },
        "CreateBadgeResponse": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "style": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "PatchBadgeRequest": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "style": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                }
            }
        }
    }
}